Enable REAL for Hive connector
* Make buckets, partitions and type utils aware of REAL type
* Support Parquet, ORC, Textfile, RCFILE and RC BINARY file types
* Update TupleDomainParquetPredicate such that REAL predicates
continue to be pushed down
* Update TupleDomainOrcPredicate such that REAL predicates
continue to be pushed down. For ORC, both float and double
statistics are stored in DoubleStatistics
(https://cwiki.apache.org/confluence/display/Hive/LanguageManual+ORC)

+++ presto-hive/src/main/java/com/facebook/presto/hive/HiveType.java
@@ -45,6 +45,7 @@ import static com.facebook.presto.spi.type.DateType.DATE;
 import static com.facebook.presto.spi.type.DecimalType.createDecimalType;
 import static com.facebook.presto.spi.type.DoubleType.DOUBLE;
 import static com.facebook.presto.spi.type.IntegerType.INTEGER;
+import static com.facebook.presto.spi.type.RealType.REAL;
 import static com.facebook.presto.spi.type.SmallintType.SMALLINT;
 import static com.facebook.presto.spi.type.TimestampType.TIMESTAMP;
 import static com.facebook.presto.spi.type.TinyintType.TINYINT;
@@ -266,7 +267,7 @@ public final class HiveType
             case LONG:
                 return BIGINT;
             case FLOAT:
-                return DOUBLE;
+                return REAL;
             case DOUBLE:
                 return DOUBLE;
             case STRING:
